Check if field is pointer before checking if field is an array
Considering the following field definition:

```c
char *text[AUDIT_MAX_ARGS];
```

Previously, `TokenParser._parse_field` would first check if a
field is an array, and then if a field is a pointer. This would
result in the field being interpreted as a pointer to an array.
However, this should have been interpreted as an array of pointers.

This issue became prevalent in `Compiler.gen_read_block`, that would
first check if a field's type is Pointer and only then if the field's
type is an Array of Pointers. As far as I can tell, this condition
cannot be reached with the current implementation of `_parse_field`.

This change performs the check if the field is a pointer before the
check if the field is an array. This would result in the field
being interpreted as an Array of Pointers.