Add support for two-dimensional arrays
A field may be defined as a two-dimensional array (array of arrays)
in case of an array of strings, which are actually character arrays.

The proposed changes allow the parser to read field names with a
secondary set of square brackets with an optional length. Furthermore,
the compiler is updated in order to actually read the child arrays.