Unpack enum when generating fields.
If an enum was a bitfield, after another bitfield, there was an error where the compiler thought there were more bitfields coming.
The issue that occurred was that that the current offset got updated, while it shouldn't have.
Afterward, there was a dynamic sized array, which tried to correct the error by seeking to the new offset, from the start of the stream.